core/src/main/java/io/kestra/plugin/core/InfluxDB/InfluxDBPlugin.java
package io.kestra.plugin.core.InfluxDB;

public class InfluxDBPlugin {
    private String url;
    private String database;
    private String username;
    private String password;

    // Default constructor
    public InfluxDBPlugin() {}

    // Method to connect to InfluxDB
    public boolean connect(String url, String database, String username, String password) {
        // Validate inputs
        if (url == null || url.isEmpty() || database == null || database.isEmpty()) {
            System.err.println("Invalid parameters: URL or Database cannot be null or empty.");
            return false;
        }

        try {
            this.url = url;
            this.database = database;
            this.username = username;
            this.password = password;

            // Simulate connection logic
            System.out.println("Connecting to InfluxDB with URL: " + url);
            // Assume connection is successful for valid parameters
            return true;
        } catch (Exception e) {
            System.err.println("Failed to connect to InfluxDB: " + e.getMessage());
            return false;
        }
    }

    // Method to insert data into InfluxDB
    public boolean insertData(String measurement, String fields) {
        if (measurement == null || measurement.isEmpty() || fields == null || fields.isEmpty()) {
            System.err.println("Invalid parameters: Measurement or Fields cannot be null or empty.");
            return false;
        }

        try {
            // Simulate data insertion logic
            System.out.println("Inserting data into measurement: " + measurement + " with fields: " + fields);
            // Assume successful insertion
            return true;
        } catch (Exception e) {
            System.err.println("Failed to insert data: " + e.getMessage());
            return false;
        }
    }
}
